refactor(project-conf): split index.ts into focused modules
